+++ Cargo.toml
@@ -13,7 +13,6 @@ wasm-opt = ["binary-install", "platforms", "tempfile"]
 
 [dependencies]
 anyhow = "1.0.52"
-binaryen = { version = "0.12.0", optional = true }
 binary-install = { version = "0.0.2", optional = true }
 cargo_metadata = "0.14.1"
 fs_extra = "1.2.0"
@@ -22,7 +21,7 @@ log = "0.4.14"
 notify = "4.0.17"
 platforms = { version = "2.0.0", optional = true }
 structopt = "0.3.25"
-tempfile = { version = "3.3.0", optional = true }
+tempfile = { version = "3.2.0", optional = true }
 walkdir = "2.3.2"
 wasm-bindgen-cli-support = "0.2.68"
